requirements.txt
asgiref==3.9.2
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
django-gis==3.2.0
celery==5.3.4
redis==4.5.5
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.5.0
django-cors-headers==4.3.1
django-filter==23.3
requests==2.32.5
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
model_bakery==1.11.0
factory-boy==3.3.0
drf-yasg==4.15.0
django-redis==5.1.0
